Packaging and portion Refactoring
